pyproject.toml
[tool.poetry]
name = "kami-reports"
version = "0.2.0"
description = ""
authors = ["Maicon de Menezes <maicon@kamico.com.br>"]
readme = "README.md"
packages = [{include = "kami_reports"}]

[tool.poetry.scripts]
kami_reports = "service:main"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.0"
kami-logging = "^0.2.1"
kami-messenger = "^0.2.0"
kami-gdrive = "^0.1.3"
pandas = "^2.0.2"
openpyxl = "^3.1.2"
xlsxwriter = "^3.1.2"
jinja2 = "^3.1.2"
pymysql = "^1.0.3"
sqlalchemy = "^2.0.15"
numerize = "^0.12"
unidecode = "^1.3.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.1.0"
blue = "^0.9.1"
isort = "^5.12.0"
taskipy = "^1.11.0"
ipython = "^8.14.0"
ipykernel = "^6.23.1"
notebook = "^6.5.4"
nbformat = "^5.9.0"


[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.1.15"
mkdocstrings = "^0.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
lint-review = "blue --check --diff . && isort --check --diff ."
lint-fix = "blue . && isort ."
pre_test = "task lint-review"
test = "pytest -s -x --cov=kami_messenger -vv -rs"
post_test = "coverage html"
docs = "mkdocs serve"